Refactor view order setting to coordinate frame drawer view order with node view order

+++ src/main/kotlin/solve/scene/view/drawing/FrameElement.kt
@@ -3,7 +3,7 @@ package solve.scene.view.drawing
 import javafx.scene.paint.Color
 import solve.scene.model.Point
 
-abstract class FrameElement(initialViewOrder: Short) : Comparable<FrameElement> {
+abstract class FrameElement(initialViewOrder: Int) : Comparable<FrameElement> {
     var viewOrder = initialViewOrder
         protected set
