+++ src/main/kotlin/solve/scene/view/drawing/ImageFrameElement.kt
@@ -7,7 +7,7 @@ import solve.scene.model.PointPairs
 import solve.scene.model.Size
 import solve.utils.ceilToInt
 
-class ImageFrameElement(viewOrder: Short, private val image: Image) : FrameElement(viewOrder) {
+class ImageFrameElement(viewOrder: Int, private val image: Image) : FrameElement(viewOrder) {
     override val points = PointPairs.getPairs(Size(image.width.ceilToInt(), image.height.ceilToInt()))
     override fun getColor(point: Point): Color = image.pixelReader.getColor(point.x.toInt(), point.y.toInt())
 }
